add: MessagesManager
