+++ static/index.html
@@ -4,6 +4,7 @@
     <meta charset="UTF-8" />
     <title>JSON Command Sender</title>
     <link rel="stylesheet" href="styles.css" />
+    <link rel="stylesheet" href="messages.css" />
     <link rel="stylesheet" href="https://cdnjs.cloudflare.com/ajax/libs/font-awesome/6.5.1/css/all.min.css" integrity="sha512-DTOQO9RWCH3ppGqcWaEA1BIZOC6xxalwEsw9c2QQeAIftl+Vegovlnee1c9QX4TctnWMn13TZye+giMm8e2LwA==" crossorigin="anonymous" referrerpolicy="no-referrer" />
     <link rel="icon" href="favicon.svg" type="image/svg+xml" />
   </head>
@@ -184,7 +185,8 @@
     </div>
 
     <script type="module" src="js/main.js"></script>
-  <script type="module" src="js/command-options.js"></script>
-  <script type="module" src="js/toast.js"></script>
+    <script type="module" src="js/command-options.js"></script>
+    <script type="module" src="js/messages.js"></script>
+    <script type="module" src="js/ui.js"></script>
   </body>
 </html>
