+++ static/messages.css
@@ -1,0 +1,200 @@
+/* ============================= */
+/* MESSAGE PANE STYLES */
+/* ============================= */
+
+#messagesContainer {
+    margin-top: 10px;
+    display: flex;
+    flex-direction: column;
+    height: 300px;
+    min-height: 200px;
+    max-height: 400px;
+    flex-shrink: 0;
+    max-width: 65vw;
+    max-width: 95vw;
+}
+
+pre#messagesDisplay {
+    flex-grow: 1;
+    overflow-y: auto;
+    overflow-x: auto;
+    background-color: #f5f5f5;
+    border: 1px solid #ccc;
+    padding: 10px;
+    margin-top: 5px;
+    font-family: monospace;
+    font-size: 0.9em;
+    white-space: pre;
+    word-break: normal;
+    min-height: 0;
+}
+
+/* Message styling - base properties */
+.message-line {
+    display: block;
+    padding: 2px 4px;
+    margin-bottom: 2px;
+    white-space: nowrap;
+    border-radius: 3px;
+    width: 100%;
+    box-sizing: border-box;
+    min-width: max-content;
+}
+
+/* Legend styling - base properties */
+.message-sent-legend, 
+.message-received-legend, 
+.message-system_info-legend, 
+.message-system_warn-legend, 
+.message-system_error-legend {
+    padding: 2px 4px;
+    border-radius: 3px;
+    font-size: 0.85em;
+}
+
+/* Sent message styling (both for messages and legend) */
+.message-sent, .message-sent_text, .message-sent-legend {
+    background-color: #a0e0fd;
+    color: #000000;
+}
+
+/* Received message styling (both for messages and legend) */
+.message-received, .message-received-legend {
+    background-color: #a4fca3;
+    color: #000000;
+}
+
+/* System message styling tiers */
+.message-system_info, .message-system_info-legend {
+    background-color: #e0e0e0;
+    color: #3a3a3a;
+}
+
+.message-system_warn, .message-system_warn-legend {
+    background-color: #fff0a3;
+    color: #856404;
+}
+
+.message-system_error, .message-system_error-legend {
+    background-color: #ffcccb;
+    color: #721c24;
+    font-weight: bold;
+}
+
+#sortMessagesButton {
+    padding: 2px 5px;
+    font-size: 0.8em;
+    margin-left: auto;
+}
+
+/* ============================= */
+/* TOAST NOTIFICATION STYLES */
+/* ============================= */
+
+#toast-container {
+    position: fixed;
+    top: 20px;
+    left: 50%;
+    transform: translateX(-50%);
+    z-index: 2000;
+    display: flex;
+    flex-direction: column;
+    align-items: center;
+    gap: 10px;
+    pointer-events: none; /* Allow clicks to pass through */
+}
+
+.toast {
+    background: #fff;
+    border-left: 5px solid #333;
+    border-radius: 8px;
+    box-shadow: 0 4px 12px rgba(0, 0, 0, 0.15);
+    padding: 15px 20px;
+    display: flex;
+    flex-direction: column;
+    align-items: flex-start;
+    gap: 5px;
+    opacity: 0;
+    transform: translateY(-20px);
+    transition: all 0.3s ease-in-out;
+    max-width: 350px;
+    min-width: 300px;
+    font-size: 0.95em;
+    cursor: pointer;
+    pointer-events: auto; /* Toasts are clickable */
+}
+
+.toast.show {
+    opacity: 1;
+    transform: translateY(0);
+}
+
+.toast-content {
+    display: flex;
+    align-items: center;
+    gap: 10px;
+    width: 100%;
+}
+
+.toast p {
+    margin: 0;
+    color: #333;
+    line-height: 1.4;
+}
+
+.toast i {
+    font-size: 1.5em;
+}
+
+.toast-dismiss-note {
+    font-size: 0.8em;
+    color: #888;
+    align-self: flex-end;
+}
+
+/* Toast types */
+.toast-info {
+    border-left-color: #007bff;
+}
+.toast-info i {
+    color: #007bff;
+}
+
+.toast-success {
+    border-left-color: #28a745;
+}
+.toast-success i {
+    color: #28a745;
+}
+
+.toast-warn {
+    border-left-color: #ffc107;
+}
+.toast-warn i {
+    color: #ffc107;
+}
+
+.toast-error {
+    border-left-color: #dc3545;
+}
+.toast-error i {
+    color: #dc3545;
+}
+
+/* ============================= */
+/* RESPONSIVE ADJUSTMENTS */
+/* ============================= */
+
+@media (max-width: 768px) {
+    #messagesContainer {
+        max-width: 95vw;
+    }
+    
+    #toast-container {
+        right: 10px;
+        left: 10px;
+        top: 10px;
+        width: auto;
+        align-items: center;
+    }
+}
